src/requirements.txt
# ─────────────────────────────────────────
# Django Core
# ─────────────────────────────────────────
Django==5.1.4
gunicorn==21.2.0
whitenoise==6.7.0

# ─────────────────────────────────────────
# Database
# ─────────────────────────────────────────
psycopg2-binary==2.9.9
django-environ==0.11.2

# ─────────────────────────────────────────
# Redis & Celery
# ─────────────────────────────────────────
redis==5.0.8
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
channels==4.1.0
channels-redis==4.2.0

# ─────────────────────────────────────────
# REST API
# ─────────────────────────────────────────
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.29.0
django-debug-toolbar==4.4.6
rcssmin==1.1.2
rjsmin==1.2.2

# ─────────────────────────────────────────
# 인증 / 권한
# ─────────────────────────────────────────
django-allauth==65.3.0

# ─────────────────────────────────────────
# 파일 / 미디어
# ─────────────────────────────────────────
Pillow==10.4.0
django-storages==1.14.4

# ─────────────────────────────────────────
# PDF 생성
# ─────────────────────────────────────────
reportlab==4.2.2
weasyprint==62.3
pydyf==0.11.0

# ─────────────────────────────────────────
# GIS / 지도
# ─────────────────────────────────────────
requests==2.32.3

# ─────────────────────────────────────────
# 데이터 처리
# ─────────────────────────────────────────
pandas==2.2.2
openpyxl==3.1.5
numpy==2.0.2

# ─────────────────────────────────────────
# 네트워크 / SNMP
# ─────────────────────────────────────────
pysnmp==6.2.5
paramiko==3.4.1
netaddr==1.3.0
python-docx==1.2.0
anthropic>=0.86.0

# ─────────────────────────────────────────
# QR코드
# ─────────────────────────────────────────
qrcode[pil]==7.4.2

# ─────────────────────────────────────────
# OCR / AI 연동
# ─────────────────────────────────────────
pytesseract==0.3.13
httpx==0.27.2

# ─────────────────────────────────────────
# 전자서명
# ─────────────────────────────────────────
cryptography==43.0.1

# ─────────────────────────────────────────
# 유틸
# ─────────────────────────────────────────
pyotp==2.9.0
django-ratelimit==4.1.0  # import as django_ratelimit
psutil==5.9.8
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2
django-redis==5.4.0
python-pptx
